Fix Mypy Contract Alignment & Penny Standard Hardening
- Enforce Penny Standard (int) in CentralBank and TransactionProcessor
- Align Protocols: IGovernmentPolicy, IMarket, IInventoryHandler
- Fix Liskov violations in LaborMarket, InventoryManager, CentralBank
- Resolve abstract class instantiation errors in Initializer
- Clean up duplicate Exception definitions in Finance API

+++ modules/finance/central_bank/service.py
@@ -54,7 +54,7 @@ class CentralBankService:
         self,
         treasury_service: ITreasuryService,
         operation_type: Literal["purchase", "sale"],
-        target_cash_amount: float,
+        target_cash_amount: int,
         currency: CurrencyCode = DEFAULT_CURRENCY
     ) -> OpenMarketOperationResultDTO:
         """
